// src/Models/IStyles.ts
export default interface IStyles {
    id: string;
    name: string;
    primaryColor: string;
    secondaryColor: string;
    backgroundColor: string;
    categoriesBackgroundColor: string;
    tagsBackgroundColor: string;
    categoriesColor: string;
    tagsColor: string;
    textColor: string;
    bannerTextBold: boolean;
    bannerTextShadow: boolean;
    bannerTextAnimation: boolean;
    bannerTextFontFamily: string;
    bannerTextShadowColor: string;
    linkColor: string;
    hoverColor: string;
    headingFont: string;
    bodyFont: string;
}